+++ src/pages/MyPage/components/SideNavigation/SideNavigation.tsx
@@ -1,34 +1,22 @@
-import { useEffect, useState } from "react";
+import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import { useUserActions } from "@/recoil/actions/auth";
-import user from "@/api/user";
 import Text from "@/components/Text";
 import UserInfo from "@/components/UserInfo";
 import { userState } from "@/recoil/state/authState";
-import { UserInfoType } from "@/components/UserInfo/UserInfo";
+import useUserInfo from "../../hooks/useUserInfo";
 import * as S from "./SideNavigation.style";
 
 const SideNavigation = () => {
-  const [userInfo, setUserInfo] = useState<UserInfoType | null>(null);
-  const [loading, setLoading] = useState(true);
   const userActions = useUserActions();
   const myUserId = useRecoilValue(userState);
+  const [userInfo, loading, userError] = useUserInfo(myUserId);
   const navigate = useNavigate();
 
   useEffect(() => {
-    const fetchUserInfo = async () => {
-      try {
-        if (!myUserId) navigate("/");
-        const data = myUserId && (await user.getUserInfo(Number(myUserId)));
-        setUserInfo(data);
-      } catch (error) {
-        navigate("/");
-      }
-    };
-    fetchUserInfo();
-    setLoading(false);
-  }, [myUserId]);
+    if (userError) navigate("/");
+  }, [userError]);
 
   if (loading || !userInfo) return <div>Loading</div>;
   return (
@@ -44,10 +32,10 @@ const SideNavigation = () => {
         <Text variant="h6" textStyle={S.MenuTextStyle}>
           라이딩 관리
           <S.StyledLink to="/mypage/riding">
-            <Text variant="subtitle1">라이딩 현황</Text>
+            <Text variant="subtitle1">라이딩 내역</Text>
           </S.StyledLink>
           <S.StyledLink to="/mypage/rating">
-            <Text variant="subtitle1">라이딩 평가</Text>
+            <Text variant="subtitle1">노쇼 체크 / 평가</Text>
           </S.StyledLink>
         </Text>
         <Text variant="h6" textStyle={S.MenuTextStyle}>
